Unify logging implementations and remove simplelog (#10686)
### Problem

There are several inconsistencies and problems with pants logging at the moment. One is that (cf. https://github.com/pantsbuild/pants/issues/10646 ) the `logging` Rust crate currently uses two separate and slightly-different methods to output logs - a custom log string if pants is displaying stderr logs when the dynamic UI is active; and a custom wrapper around the `simplelog::WriteLogger` when writing logs to a file or displaying stderr logs when the dynamic UI is inactive.

Another issue is that Rust logging can sometimes be the source of deadlocks. Many pants log statements log the string representation of a `Value` type, which represents a Python object and calls into Python code to determine its string representation, which requires grabbing the Python GIL. `WriteLogger` uses a lock internally around its handle to a `Write`-implementing value, which can interfere with grabbing the GIL in the `Display` implementation of `Value`, causing deadlocks if we display log messages frequently enough.

### Solution

This commit fixes both the inconsistency issue and the deadlocking issue by unifying the logging implementations in `logger.rs` and removing our `MaybeWriteLogger` wrapper around `simplelog::WriteLogger` entirely. Now, the custom log string built for the dynamic UI stderr logging case is the only log format string in use; and since we compute this string before grabbing the lock to a file (and remove the lock entirely for stderr), we avoid the deadlocks we were seeing.