remoting: output directories in action results reference trees (#10691)
### Problem

The `tree_digest` field of `OutputDirectory` protos embedded in an `ActionResult` is a digest pointing to a `Tree` and not a `Directory`. The RE client, however, wrongly assumes that it is a `Directory` message. Thus, Pants is unable to decode `ActionResult` protos that have output directory responses.

>  // The digest of the encoded
  // [Tree][build.bazel.remote.execution.v2.Tree] proto containing the
  // directory's contents.
  Digest tree_digest = 3;

[See the Remote Execution spec for more details.](https://github.com/bazelbuild/remote-apis/blob/5a7b1a472165a0ea47b9060089855385fe351193/build/bazel/remote/execution/v2/remote_execution.proto#L1097-L1100)

### Solution

Rewrite the parsing of action results so that the client decodes the `Tree` referenced in the `ActionResult` and hashes the embedded root `Directory` for the applicable `OutputDirectory`.

### Result

Tests pass.