Fix `test --debug` to be hermetic, but keep `run` and `repl` non-hermetic (#10701)
## Problem

See https://github.com/pantsbuild/pants/pull/10668 for the original implementation, which ended up getting reverted.

As pointed out there and in https://github.com/pantsbuild/pants/issues/10644, there's an argument to keeping `run` and `repl` non-hermetic, i.e. using the env for the parent `./pants` process. Neither of those goals are cached. REPL is designed for experimentation. Run's analog is `binary`; when producing a `binary` and then running it, the `binary` would end up having the env of the host platform, rather than what Pants sets.

However, we do want `test --debug` to be hermetic so that `test` and `test --debug` behave the same.

## Solution

Add the argument `hermetic_env: bool` to the `InteractiveProcess` constructor, which controls whether the env gets stripped or not. We default this to `True` as, generally, hermiticity is a good default.

`run.py` and `repl.py` ensure that `hermetic_env` is always set to False.

### Nuance: env var may be still be overridden

If the `InteractiveProcess` includes an env var that is already set in the parent process's env, then we will override the env var with Pants's value.

Tangibly, the Python implementations for `repl` and `run` will override the `$PATH` to whatever is set via `--pex-executable-search-paths`.

[ci skip-build-wheels]