Fix invalid shebang for `BinaryPaths` script (#10710)
### Problem

The code in https://github.com/pantsbuild/pants/blob/master/src/python/pants/engine/process.py, which uses a shell script to probe where Python interpreters are located, puts a newline in front of the shebang which can error out when remote execution environments try to execute the script. Unix/Linux requires that the shebang be the very first bytes in the file.

### Solution

Strip whitespace off the front of the script. This removes the newline just after the `"""` marker.

### Result

Remote execution in Buildbarn should pass.

[ci skip-rust]
[ci skip-build-wheels]