Improve RuleGraph convergence time slightly (#10700)
### Problem

One of the open issues in https://www.pantsbuild.org/v2.0/docs/internal-rules-architecture was that we weren't able to prune based on transitive information.

### Solution

Track whether the `in_set` of a node is transitively minimal, which allows us to prune nodes that don't consume a provided `Param` sooner. Additionally, document some of the other things I've explored recently, and add a minimal testcase that goes into an infinite loop without the `splits` special case.

### Result

This does not fix #10683 unfortunately. Still on it.