Fix retrieval of bytes from artifacts in StreamingWorkunitHandler (#10698)
### Problem

The `digests_to_bytes` method, used to allow the StreamingWorkunitHandler to get the raw bytes associated with a `Digest` item on a workunit, works correctly for the `stdout_digest` and `stderr_digest` items. These effectively represent nameless files whose content is the stdout and stderr stream of a process, and the `Digest`s attached to the workunit yield the bytes the user expects. 

However, `digests_to_bytes` fails for other `Digest`s provided as workunit artifacts, because these generally come from a `Snapshot` that was the output of some `@rule`. The `Digest` contained within a `Snapshot` is generally a directory containing a single named file; trying to transform this kind of `Digest` into bytes with `digests_to_bytes` fails at the `store` level and produces a Python-level error.

### Solution

This commit adds a new method on `StreamingWorkunitContext` called `snapshots_to_file_contents`. This method can convert a sequence of `Snapshot`s into a sequence of sequences of `FileContent` representing the files contained within that snapshot, which in turn contain the byte values of those files. As well, the various python code flows that pass `Digest`s around to be used on the `artifacts` method have been modified to now pass around `Snapshot`s. 

A plugin that receives streaming workunits can now use the `snapshots_to_file_contents` context method to correctly retrieve the bytes of artifacts on a workunit, which are now more-correctly represented as a `Snapshot` rather than a `Digest`. And since an artifact is now a `Snapshot`, in principle any number of discrete files can now constitute a single artifact.