Add `RuleRunner.get_target()` (#10714)
We often request a target in tests for some address. This is common enough to be worth factoring up.

We also port `run_setup_py_test.py` to use Pytest style, and to use this new util. In addition, we make some small improvements to `run_setup_py.py`. This is prework for https://github.com/pantsbuild/pants/issues/10633.