Improve Pylint to use `--python-setup-resolve-all-constraints` (#10726)
We want this optimization so that we can use one single resolve for the whole repo, and so that the single resolve gets reused across `test`, `run`, and `repl`.

However, unlike everything else, Pylint only considers direct dependencies, rather than transitive dependencies. We want to keep this because it results in far fewer cache invalidations, which is relevant if `--python-setup-resolve-all-constraints` is not in play. It's also important that we only consider direct dependencies because we do not consider transitive dependencies when we are partitioning targets based on their interpreter constraints.

Will close #10370.

[ci skip-rust]
[ci skip-build-wheels]