Support adding arbitrary env vars to test processes (#10715)
### Problem

cf. https://github.com/pantsbuild/pants/issues/10644 , we would like to be able to specify additional environment variables that can be made available to test processes, both explicitly specified and with a "shorthand" option that can read the value of an environment variable from Pants' own execution environment. 

### Solution

This commit introduces a new type `PantsEnvironment` that contains a copy of the environment variables a pants process was invoked with. Since this type contains an exact copy of the environment, we expect that its values might change spuriously, so we cannot use it directly as the input to rules without risking frequent needless cache invalidations. Instead, we mark this type
as side-effecting so it can only be included in a non-cacheable rule, and add some logic to prune the environment to only the variables explicitly requested by a (new) `--test-extra-env-vars` option.

`PantsEnvironment` is a general mechanism that can be used in any rule that needs to be aware of pants' own environment. The remaining work in this commit is to hook up the pruned result from `PantsEnvironment` into a new `TestExtraEnv` mechanism, and wire that up to the test process execution framework.