Disambiguate binary and lambda output. (#10735)
Previously we named them after the target name, which might
not be unique across the repo. Now we use the entire target
address. The new name is `path.to.target.dir/target_name.pex/zip`.
This seems like a good compromise between completely flat
and overstructured. It also preserves the existing filenames,
for less confusion.

[ci skip-rust]