Support binary dependencies in tests.
A test can now depend on a binary target (python or otherwise)
and the built binary will be present in the test's chroot.

Note that this supports a user's immediate use case, but doesn't
address some wider questions of what it means to depend on a binary
target. For example, if a binary depends on another binary, should
that second binary be embedded as a resource inside the first one?
But until we figure out what the general principle might be,
this is useful in a real-world case.

[ci skip-rust]

[ci skip-build-wheels]