Fix MyPy to work by default with Python 3.6+ code and Black with 3.8+ code (#10750)
### Problem

MyPy and Black both use typed-ast for parsing, just like we use it for dependency inference. Typed-ast understands Py27 and 34-37, but not Py38. For Py38, you instead must run with a Python 3.8 interpreter (which also understands 27 and 34-37).

This meant that our default Black and MyPy installations would break if encountering Py38 syntax like the walrus operator. You had to explicitly set `--black-interpreter-constrains` and `--mypy-interpreter-constraints`, which is not at all obvious.

### Solution

Detect if the user has constraints for 3.8+. If so, use the constraints `>=3.8` for the tools. Otherwise, use the tool's default constraints.

We check that the constraints do not include anything less than 3.8, such as 3.7. Why? If they need to support 3.7, then a user won't be using 3.8 syntax. We don't want to constraint to `>=3.8` because this would mean that they must have 3.8 on their machine, even though 3.7 would do just fine.

To facilitate this, we change `PexInterpreterConstraints` to store parsed `Requirement` objects.

#### Also fixes MyPy handing of Python 3.6 and Python 3.7

If the user leaves off `python_version`, MyPy defaults to the version of the interpreter used to run. MyPy also tells the AST parser to target this specific `python_version`. So, if you're running with our minimum requirement of Python 3.5, it will fail to parse any Python 3.6+ code like f-strings, unless you specifically set `python_version` in `mypy.ini`.

We apply the same fix as above so that MyPy now also works "out of the box" with Python 3.6-only and 3.7-only codebases. If users set `python_version` in `mypy.ini`, MyPy will respect this - we only influence the default behavior.

### Result

Toolchain's Python 3.8 codebase now works "out of the box" with Pants, i.e. without extra settings. Pants's codebase, which uses Py36, also works out of the box now, even when Py35 is available.

[ci skip-rust]
[ci skip-build-wheels]