Add support to MyPy for third-party requirements and stubs (PEP 561) (#10753)
Similar to with v1, we go through some hoops with our own launcher script that monkey patches the module discovery to work with Pex. This is explained in the code's comments.

If users use a constraints.txt file, there will be no meaningful performance hit thanks to `--python-setup-resolve-all-constraints`. Otherwise, this will result in much slower perf due to resolving requirements.

[ci skip-rust]
[ci skip-build-wheels]