Add `runtime_binary_dependencies` field to `python_tests` target (#10756)
This improves on the mechanism introduced in https://github.com/pantsbuild/pants/pull/10744. It's not consistent that putting a `python_binary` in the `dependencies` field would result in actually building a binary during runtime. That only happened with `python_tests`, which is confusing.

It can also be surprising when Pants does this for you automatically, given how heavy-weight building a binary can be (e.g. resolving requirements). For example, as a result of #10744, all of our own integration tests would end up building `local_pants.pex`, which we did not want.

Now, you must explicitly enumerate every binary that you want built in the field `runtime_binary_dependencies`.

[ci skip-build-wheels]
[ci skip-rust]