Add support to MyPy for first-party plugins (#10755)
Similar to with Pylint plugins, we introduce a new target type `mypy_source_plugin`, along with a new option `--mypy-source-plugins`. 

The new target type is not strictly necessary; you could use a `python_library`. But, the new target type a) provides clearer modeling of what's going on, e.g. that this is a "root" rather than a typical library; and b) allows us to add instructions via `./pants target-types --details=mypy_source_plugin`.
 
[ci skip-build-wheels]
[ci skip-rust]