Support volatile processes. (#10768)
Support uncacheable processes.

This just wraps up the current UUID hack behind a type to prevent the
surface area of the hack from spreading. If we eventually plumb a
cacheability flag through Process to the Rust side we can still do so 
with a small change to the current UUID env var hack and a controlled
deprecation of the UncacheableProcess type in favor of the Process flag
should it come to exist.