remoting: move target_platform property into environment variable (#10713)
### Problem

Pants currently sets a `target_platform` property automatically in the platform properties of remote execution requests. This property does not use the name or semantics of what the REv2 protocol suggests (i.e., the `OSFamily` property). 

Moreover, the fact that this property can be set to multiple values in the same run (e.g. `linux` and `none`) interferes with setting up workers on the server side, especially for Buildbarn which allocates work based on an exact match to platform properties and Buildgrid which just fails for a worker to pick up the work if the second unconfigured property is seen.

See discussion in https://github.com/pantsbuild/pants/issues/10010 for more information.

### Solution

Move the `target_platform` property on remote execution requests into an environment variable so it does not implicate platform property semantics on the server but still is in the cache key.

### Result

Updated tests pass.