Run internal only pexes via discovered python.
Since shebangs record information about the host system it is incorrect
to ever rely on these when present in a binary stored in the CAS. A
shebang that worked in the past can fail in the future even though a viable
alternative interpreter may be discoverable.

As such, convert all local internal-only PEX creation and execution to
using a freshly discovered interpreter.

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]