Signal handling fixes (#10758)
### Problem

Ideally, pants should exit promptly when a user hits Ctrl-C in the terminal, regardless of how it is configured to run or what it is running. cf. https://github.com/pantsbuild/pants/issues/10051 and other general observations made while using pants, there are still a few cases where pants does not respond quickly to a Ctrl-C and hangs or otherwise takes a while to quit. This commit is an attempt to fix these cases.

### Solution

In the `--debug` case, the problem turned out to be that we were explicitly disabling SIGINT handling in the case where pants was running an `InteractiveProcess`. For the non-pantsd case, this is desirable - we use the `InteractiveProcess` abstraction for `test --debug` and a few other cases where we legitimately want a pants subprocess to take over the terminal and do its own signal handling. But when pantsd is running, we don't want to explicitly disable SIGINT handling, because that means that pantsd will ignore a signal from its own client process. This commit solves the problem by adding a flag to `SignalHandler` to indicate whether this is a pantsd run, and doesn't ignore SIGINT if it is.

Some other instances of pants ignoring prompt signal handling are caused by rust code not heeding the Python interrupt state triggered by SIGINT, resulting in pants not quitting until execution returns to the Python main thread. This commit works around this by having the Python signal handling infrastructure set a flag on `ExceptionSink` once it has handled a signal that would cause pants to terminate. A python helper function is passed into the main execution loop in `scheduler.rs` that checks for the presence of this flag, and breaks the loop to terminate the engine if it finds that the flag is set. 

### Result

Ctrl-C should now work in the `test --debug` case with pantsd, and work in the `--no-pantsd` case when any kind of test is running. It still does seem to take more than one hit of Ctrl-C to fully quit pants in some cases, though, which will be addressed in followup commits.