@rules consume params to prevent some types of ambiguity (#10764)
### Problem

Codegen backends are currently disabled because enabling them introduced new `@rule` paths/cycles in the RuleGraph that took minutes to resolve pre #10645, and post #10645 failed with an error. Additionally, "transitive excludes" (#10661) were disabled due to similar issues in order to land #10645.

### Solution

* Change the semantics around which `Param`s are in scope to remove them from scope in a `@rule`'s dependencies when they are directly consumed as a positional argument by a `@rule`. This required no changes to any `@rules` in the pantsbuild/pants repo. See the new section in `README.md` for more information.
* Fix the cycle-breaking `TODO` in `monomorphize` by removing "re-joining" of nodes (they only split, and even if an identical copy is created, they don't merge) and replacing the "splits" workaround with recording which nodes we suspect to be ambiguous, and marking them deleted as ambiguous if they split again in the same way.
* Move error rendering entirely to the `prune_edges` phase, using the deletion information from the earlier phases.

### Result

Fixes #10683 and fixes #10661.

The semantics changes might be visible to `@rule` authors: they would exhibit as a `@rule` failing because it didn't have a legal source of a `Param`: followup work to improve errors will be necessary.