Simplify --print-exception-stacktrace logic (#10786)
Currently, we're using the value of the `--print-exception-stacktrace` option to set a class var on `ExceptionSink` early in the pants initialization process. The `ExceptionFormatter` class in the Python logging initialization code uses this value as its source of truth for whether or not to emit stacktrace output. There's also one code path in `ExceptionSink` that uses this value, namely the code that responds to SIGTERM and SIGQUIT.

I'm actually not sure if the SIGTERM/SIGQUIT handling code is actually doing anything right now - it seems like if those signals are actually sent to the pants process, the failure will actually happen in Rust engine code, and never touch that code path. In any case, since SIGQUIT and SIGTERM are only going to be sent to pants very deliberately, it should be fine to default to always printing a stack trace in that eventuality.

This means that the code around `ExceptionSink` that pays attention to this variable can be removed entirely, and the global `--print-exception-stacktrace` option can be the only source of truth for this value.