Remove reset_interactive_output_stream (#10791)
### Problem

`ExceptionSink.reset_interactive_output_stream` is a legacy codepath that touches some class variables in `ExceptionSink` as well as pants initialization code in both the pantsd and no-pantsd cases. The only case where this code is relevant is in the custom singal handler for SIGUSR2, where the python standard library `faulthandler` module was being used to optionally print a Python stack trace when that signal was received (which is also an undocumented piece of functionality).

### Solution

This commit removes that code path entirely, which allows us to simplify `ExceptionSink` and remove a dependency on `ExceptionSink` in several places. 

cf. https://github.com/pantsbuild/pants/issues/10792 , we may want to re-introduce the functionality to send pants some kind of signal to output a live stack trace. However, if we do this, we should re-implement the feature in a way that is aware of the Rust engine (and useful for debugging Rust as well as Python issues), and make sure that we document it.