Un-revert Add cache_failures option to Process type (#10796)
After https://github.com/pantsbuild/pants/pull/10433 (Add cache_failures option to Process type) was merged, it was noted that that change caused failing tests to repeat ( https://github.com/pantsbuild/pants/issues/10129#issuecomment-667774204 ), and the original commit was reverted. I spent some time digging into the `graph` crate to try to figure out the cause of the problem, but was unsuccessful. 

However, as of today, we found that we could no longer reproduce the original bug. There have been some significant changes to `rule_graph` and several other deep parts of the Rust engine's codebase since the original revert, so it's very possible that the condition that originally triggered the bug is no longer relevant. Consequently, this commit restores the original commit.