Re-export `QueryRule` from `pants.testutil.rule_runner` (#10806)
While `QueryRule` is indeed used in production, it's very rare to be used and we do not expect plugin authors to interact with the mechanism, except for when writing tests using `RuleRunner`. 

So, to simplify the explanation, we present "Query rule" as primarily existing for the sake of tests due to needing to use the Synchronous API: https://www.pantsbuild.org/docs/rules-api-testing#setting-up-the-rulerunner

This change reinforces that message that `QueryRule` has a strong link to testing. It also reduces the number of imports.

[ci skip-rust]
[ci skip-build-wheels]