Simplify logging initialization (#10812)
This PR attempts to simplify the code paths around log initialization in the several places in the pants codebase where we do so. It also removes the `_LoggerStream` replacement for `sys.stdout` and `sys.stderr`, since code in pantsd should be using the Python logging abstraction rather than trying to write to stdout/stderr and then having that be directed to the log.