Fix bug in python binary running.
Previously we assumed that the entry point, if explicit,
was a module. However it may be module:file. This change
delegates interpreting the entry point string to Pex itself.

It also addresses a TODO in computing the implicit entry
point from sources. Now we just do it by path manipulation,
without snapshotting, as we now have the ability to do so.

[ci skip-rust]

[ci skip-build-wheels]