Internally stop ignoring missing imports with MyPy (#10828)
Now we're explicit about every dependency that is missing requirements. This is generally a best practice, and it better dogfoods that MyPy requirements are indeed working.

We add `pytest` to our `requirements.txt` so that Pants can properly infer a dep on Pytest. This is an awkward edge though that it duplicates with `--pytest-version`. This is a known issue.

[ci skip-build-wheels]
[ci skip-rust]