Fix `setup_py.with_binaries()` to automatically inject `dependencies` (#10829)
If you have a `python_binary` target in `setup_py().with_binaries()`, it's clear that you intend to depend on the target and for its files to be included in the result.

Right now, you must explicitly add it to the `dependencies` field too, otherwise you will have the right `setup.py` entry, but missing files.

This uses dependency injection (not inference) to automatically inject the dependencies for a `python_distribution` target. This approach is more robust than special casing in `setup_py.py`. For example, this ensures that our invalidation will work properly in all contexts, and that `./pants dependencies` includes it.

[ci skip-rust]
[ci skip-build-wheels]